EPMLSTRTAI-359: Remove JavaPoetClass from JavaClass inheritance hierarchy. Move JavaFileWriter to adapter folder
