+++ src/main/java/com/epam/page/object/generator/adapter/JavaClass.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.lang.model.element.Modifier;
 
 public interface JavaClass {
+    String getPackageName();
 
     String getClassName() throws IOException;
