+++ src/test/java/com/epam/page/object/generator/PageObjectsGeneratorTest.java
@@ -12,7 +12,7 @@ import com.epam.page.object.generator.parser.JsonRuleMapper;
 import com.epam.page.object.generator.validators.ValidationContext;
 import com.epam.page.object.generator.validators.ValidationResult;
 import com.epam.page.object.generator.validators.ValidatorsStarter;
-import com.epam.page.object.generator.writer.JavaFileWriter;
+import com.epam.page.object.generator.adapter.JavaFileWriter;
 import com.google.common.collect.Lists;
 import java.util.List;
 import org.junit.Before;
